#15 refactor: domain 모듈 디렉토리 구조 수정
