+++ domain/src/main/java/com/laundreader/domain/repository/user/UserRepository.java
@@ -1,11 +1,11 @@
-package com.laundreader.domain.user.repository;
+package com.laundreader.domain.repository.user;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.laundreader.domain.user.entity.User;
-import com.laundreader.domain.user.type.UserStatus;
+import com.laundreader.domain.entity.user.User;
+import com.laundreader.domain.type.user.UserStatus;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 	Optional<User> findByEmailAndStatus(String email, UserStatus userStatus);
