+++ domain/src/main/java/com/laundreader/domain/repository/withdrawLog/WithdrawLogRepository.java
@@ -1,8 +1,8 @@
-package com.laundreader.domain.withdrawLog.repository;
+package com.laundreader.domain.repository.withdrawLog;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.laundreader.domain.withdrawLog.entity.WithdrawLog;
+import com.laundreader.domain.entity.withdrawLog.WithdrawLog;
 
 public interface WithdrawLogRepository extends JpaRepository<WithdrawLog, Long> {
 }
