+++ domain/src/main/java/com/laundreader/domain/type/user/Role.java
@@ -1,0 +1,6 @@
+package com.laundreader.domain.type.user;
+
+public enum Role {
+	USER,
+	ADMIN;
+}
