--- domain/src/main/java/com/laundreader/domain/user/type/Role.java
@@ -1,6 +1,0 @@
-package com.laundreader.domain.user.type;
-
-public enum Role {
-	USER,
-	ADMIN;
-}
