+++ user-api/src/main/java/com/laundreader/userapi/_core/security/jwt/JwtTokenProvider.java
@@ -6,7 +6,7 @@ import java.util.Date;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import com.laundreader.domain.user.entity.User;
+import com.laundreader.domain.entity.user.User;
 import com.laundreader.userapi._core.AppConstants;
 
 import io.jsonwebtoken.ExpiredJwtException;
